[dpkdir] revert *MapFileObservers() addition, fix #105
Neumond probably thoughts it would be cool to make a file observer
that reloads the VFS each time the map change, was a nice idea.
Badly, it makes the VFS reloading itself before saving a map,
losing the current work done.

Would be easier to not rely on that observer mechanism and just
make an explicit call to reload VFS when needed instead, so we
revert this part.

This VFS reload mechanism will be reintroduced in next commit.

This commit fixes #105 even if DPK VFS is not yet fully restored.

Note: this observer mechanism was also made available to plugins, so
it's not available to plugins anymore (it was not used, and it's a bad
idea, see the next commit).