[dpkdir] introducing dpkdeps.h
It was a change included in a commit by Neumond I haven't picked-up
because I disagreed with some other changes this commit introduced.

But this part is legit, it moves some DPK's DEPS file read and
version computation code to its own file. This clean-up a bit the
vfspk3 code and by the way it will be needed by future code to write
DEPS file from NetRadiant itself.

This commit is only about reorganizing code a better way.