Revert "gtk3: call ui::process() after gtk_gl_area_queue_render()"
This reverts commit 95333c891fb8e1ce33241c51ed762a4a9e1f5893.

The default value for gtk_gl_area_set_auto_render is true…
So removing the call setting it to true has no effect.

The ui::process() call was reported to improve performance
when using a multi-window layout and GL texture binds fail
because of no sharing mechanism accross GL contexts. So it's
just dark magic and we may better fix that sharing thing first.